Add more engines

+++ test_bot.py
@@ -36,6 +36,33 @@ def download_sf():
         os.chmod(f'./TEMP/sf2{file_extension}', st.st_mode | stat.S_IEXEC)
 
 
+def download_lc0():
+    response = requests.get('https://github.com/LeelaChessZero/lc0/releases/download/v0.28.2/lc0-v0.28.2-windows-cpu-dnnl.zip', allow_redirects=True)
+    with open('./TEMP/lc0_zip.zip', 'wb') as file:
+        file.write(response.content)
+    with zipfile.ZipFile('./TEMP/lc0_zip.zip', 'r') as zip_ref:
+        zip_ref.extractall('./TEMP/')
+
+
+def download_sjeng():
+    response = requests.get('https://sjeng.org/ftp/Sjeng112.zip', allow_redirects=True)
+    with open('./TEMP/sjeng_zip.zip', 'wb') as file:
+        file.write(response.content)
+    with zipfile.ZipFile('./TEMP/sjeng_zip.zip', 'r') as zip_ref:
+        zip_ref.extractall('./TEMP/')
+    shutil.copyfile(f'./TEMP/Release/Sjeng112.exe', f'./TEMP/sjeng.exe')
+
+
+if os.path.exists('TEMP'):
+    shutil.rmtree('TEMP')
+os.mkdir('TEMP')
+download_sf()
+if platform == 'win32':
+    download_lc0()
+    download_sjeng()
+lichess_bot.logger.info("Downloaded engines")
+
+
 def run_bot(CONFIG, logging_level, stockfish_path):
     lichess_bot.logger.info(lichess_bot.intro())
     li = lichess_bot.lichess.Lichess(CONFIG["token"], CONFIG["url"], lichess_bot.__version__)
@@ -124,7 +151,7 @@ def run_bot(CONFIG, logging_level, stockfish_path):
                 win = board.is_checkmate() and board.turn == chess.WHITE
                 with open('./logs/result.txt', 'w') as file:
                     file.write('1' if win else '0')
-            
+
             thr = threading.Thread(target=thread_for_test)
             thr.start()
             lichess_bot.start(li, user_profile, engine_factory, CONFIG, logging_level, None, one_game=True)
@@ -145,30 +172,77 @@ def test_sf():
     if platform != 'linux' and platform != 'win32':
         assert True
         return
-    if os.path.exists('TEMP'):
-        shutil.rmtree('TEMP')
-    os.mkdir('TEMP')
     if os.path.exists('logs'):
         shutil.rmtree('logs')
     os.mkdir('logs')
     logging_level = lichess_bot.logging.INFO  # lichess_bot.logging_level.DEBUG
     lichess_bot.logging.basicConfig(level=logging_level, filename=None, format="%(asctime)-15s: %(message)s")
     lichess_bot.enable_color_logging(debug_lvl=logging_level)
-    download_sf()
-    lichess_bot.logger.info("Downloaded SF")
     with open("./config.yml.default") as file:
         CONFIG = yaml.safe_load(file)
     CONFIG['token'] = ''
     CONFIG['engine']['dir'] = './TEMP/'
+    CONFIG['engine']['working_dir'] = './TEMP/'
     CONFIG['engine']['name'] = f'sf{file_extension}'
     CONFIG['engine']['uci_options']['Threads'] = 1
     stockfish_path = f'./TEMP/sf2{file_extension}'
     win = run_bot(CONFIG, logging_level, stockfish_path)
-    shutil.rmtree('TEMP')
-    shutil.rmtree('logs')
     lichess_bot.logger.info("Finished Testing SF")
+    assert win
+
+
+@pytest.mark.timeout(150)
+def test_lc0():
+    if platform != 'win32':
+        assert True
+        return
+    if os.path.exists('logs'):
+        shutil.rmtree('logs')
+    os.mkdir('logs')
+    logging_level = lichess_bot.logging.INFO  # lichess_bot.logging_level.DEBUG
+    lichess_bot.logging.basicConfig(level=logging_level, filename=None, format="%(asctime)-15s: %(message)s")
+    lichess_bot.enable_color_logging(debug_lvl=logging_level)
+    with open("./config.yml.default") as file:
+        CONFIG = yaml.safe_load(file)
+    CONFIG['token'] = ''
+    CONFIG['engine']['dir'] = './TEMP/'
+    CONFIG['engine']['working_dir'] = './TEMP/'
+    CONFIG['engine']['name'] = 'lc0.exe'
+    CONFIG['engine']['uci_options']['Threads'] = 1
+    CONFIG['engine']['uci_options'].pop('Hash', None)
+    CONFIG['engine']['uci_options'].pop('Move Overhead', None)
+    stockfish_path = './TEMP/sf2.exe'
+    win = run_bot(CONFIG, logging_level, stockfish_path)
+    lichess_bot.logger.info("Finished Testing LC0")
+    assert win
+
+
+@pytest.mark.timeout(150)
+def test_sjeng():
+    if platform != 'win32':
+        assert True
+        return
+    if os.path.exists('logs'):
+        shutil.rmtree('logs')
+    os.mkdir('logs')
+    logging_level = lichess_bot.logging.INFO  # lichess_bot.logging_level.DEBUG
+    lichess_bot.logging.basicConfig(level=logging_level, filename=None, format="%(asctime)-15s: %(message)s")
+    lichess_bot.enable_color_logging(debug_lvl=logging_level)
+    with open("./config.yml.default") as file:
+        CONFIG = yaml.safe_load(file)
+    CONFIG['token'] = ''
+    CONFIG['engine']['dir'] = './TEMP/'
+    CONFIG['engine']['working_dir'] = './TEMP/'
+    CONFIG['engine']['protocol'] = 'xboard'
+    CONFIG['engine']['name'] = 'sjeng.exe'
+    CONFIG['engine']['ponder'] = False
+    stockfish_path = './TEMP/sf2.exe'
+    win = run_bot(CONFIG, logging_level, stockfish_path)
+    lichess_bot.logger.info("Finished Testing Sjeng")
     assert win
 
 
 if __name__ == '__main__':
     test_sf()
+    test_lc0()
+    test_sjeng()
